Big update
This is a major update to the mle_rhythmicity toolset! We have made
substantial changes to the input/output, the math back end, and added
some functionality. mle_rhythmicity now uses definite integrals rather
than Riemann sums, making it much faster. We have also added
rhythmicity_covar, which allows for the modulation of parameters (e.g.
frequency) by any number of covariates (e.g. running speed). We hope
that people find this update useful, but if you'd like to get back to
the latest version of the old code it is in the last committed version
of the previous revision is the SHA starting with 93862ac...